+++ crates/swc_ecma_compiler/src/es2020/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod export_namespace_from;
+pub(crate) mod nullish_coalescing;
